+++ src/main/scala/com.gu.contentapi.client/parser/JsonParser.scala
@@ -76,6 +76,7 @@ object JsonParser {
     case JField("legallySensitive", JString(s)) => JField("legallySensitive", JBool(s.toBoolean))
     case JField("sensitive", JString(s)) => JField("sensitive", JBool(s.toBoolean))
     case JField("explicit", JString(s)) => JField("explicit", JBool(s.toBoolean))
+    case JField("clean", JString(s)) => JField("clean", JBool(s.toBoolean))
   }
 
   def generateJson[A <: ThriftEnum]: PartialFunction[Any, JString] = {