+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 7.8
+* Add explicit field to AssetFields
+* Add tests for audio elements
+* Add durationMinutes, durationSeconds, explicit and clean field to AudioElementField
+
 ## 7.7
 * Added the `embedType` and `html` fields to asset fields